requirements.txt
alembic==1.8.0
anyio==3.6.1
attrs==21.4.0
bcrypt==3.2.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.1
distlib==0.3.4
ecdsa==0.17.0
envtoml==0.1.2
fastapi==0.78.0
filelock==3.7.1
flake8==4.0.1
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
hyperframe==5.2.0
identify==2.5.1
idna==2.10
iniconfig==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
uvloop==0.16.0
virtualenv==20.15.1
watchfiles==0.15.0
websockets==10.3
